+++ x/stake/keeper/slash_test.go
@@ -291,10 +291,10 @@ func TestSlashWithUnbondingDelegation(t *testing.T) {
 	// just 1 bonded token burned again since that's all the validator now has
 	require.Equal(t, int64(10), oldPool.BondedTokens-newPool.BondedTokens)
 	// read updated validator
-	validator, found = keeper.GetValidatorByPubKey(ctx, pk)
-	require.True(t, found)
 	// power decreased by 1 again, validator is out of stake
-	require.Equal(t, sdk.NewRat(0), validator.GetPower())
+	// ergo validator should have been removed from the store
+	_, found = keeper.GetValidatorByPubKey(ctx, pk)
+	require.False(t, found)
 }
 
 // tests Slash at a previous height with a redelegation
@@ -387,16 +387,16 @@ func TestSlashWithRedelegation(t *testing.T) {
 	// four more bonded tokens burned
 	require.Equal(t, int64(16), oldPool.BondedTokens-newPool.BondedTokens)
 	// read updated validator
-	validator, found = keeper.GetValidatorByPubKey(ctx, pk)
-	require.True(t, found)
-	// power decreased by 4, down to 0
-	require.Equal(t, sdk.NewRat(0), validator.GetPower())
+	// validator decreased to zero power, should have been removed from the store
+	_, found = keeper.GetValidatorByPubKey(ctx, pk)
+	require.False(t, found)
 
 	// slash the validator again, by 100%
 	// no stake remains to be slashed
 	ctx = ctx.WithBlockHeight(12)
-	validator, found = keeper.GetValidatorByPubKey(ctx, pk)
-	require.True(t, found)
+	// validator no longer in the store
+	_, found = keeper.GetValidatorByPubKey(ctx, pk)
+	require.False(t, found)
 	keeper.Slash(ctx, pk, 10, 10, sdk.OneRat())
 
 	// read updating redelegation
@@ -409,10 +409,9 @@ func TestSlashWithRedelegation(t *testing.T) {
 	// no more bonded tokens burned
 	require.Equal(t, int64(16), oldPool.BondedTokens-newPool.BondedTokens)
 	// read updated validator
-	validator, found = keeper.GetValidatorByPubKey(ctx, pk)
-	require.True(t, found)
-	// power still zero
-	require.Equal(t, sdk.NewRat(0), validator.GetPower())
+	// power still zero, still not in the store
+	_, found = keeper.GetValidatorByPubKey(ctx, pk)
+	require.False(t, found)
 }
 
 // tests Slash at a previous height with both an unbonding delegation and a redelegation